chore: Move reducer hooks file
