+++ src/components/timer/TimerContainer.tsx
@@ -4,7 +4,7 @@ import { FaUndoAlt } from "react-icons/fa";
 import RunningTimer from "./RunningTimer";
 import StaticTimer from "./StaticTimer";
 import StyledIcon from "../common/StyledIcon";
-import { useAppDispatch, useAppSelector } from "../../hooks";
+import { useAppDispatch, useAppSelector } from "../../reducers/hooks";
 import { getNextMode, resetCycle, selectMode } from "../../reducers/timerReducer";
 import { playAudio } from "../../utils/media";
 import { MODE_TEXT } from "../../utils/constants";
