+++ src/reducers/hooks.ts
@@ -1,5 +1,5 @@
-import { useDispatch, useSelector, TypedUseSelectorHook } from "react-redux";
-import { RootState, AppDispatch } from "./store";
+import { useDispatch, useSelector, type TypedUseSelectorHook } from "react-redux";
+import type { AppDispatch, RootState } from "../store";
 
 type DispatchFunc = () => AppDispatch;
